services/api/requirements.txt
# Requirements for API Gateway and Signaling Server
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
websockets==14.1
asyncpg==0.30.0
redis[hiredis]==5.2.1
sqlalchemy==2.0.36
alembic==1.14.0
